requirements.txt
fastapi
uvicorn
python-multipart
pydantic
langchain
langchain-core
langchain-google-genai
pymupdf
python-dotenv
